+++ ts/navigation/AuthenticatedStackNavigator.tsx
@@ -44,24 +44,22 @@ import {
   IDPayUnsubscriptionRoutes
 } from "../features/idpay/unsubscription/navigation/navigator";
 import UnsupportedDeviceScreen from "../features/lollipop/screens/UnsupportedDeviceScreen";
+import { MessagesStackNavigator } from "../features/messages/navigation/MessagesNavigator";
+import { MESSAGES_ROUTES } from "../features/messages/navigation/routes";
+import { WalletBarcodeNavigator } from "../features/payments/barcode/navigation/navigator";
+import { PaymentsBarcodeRoutes } from "../features/payments/barcode/navigation/routes";
+import { PaymentsMethodDetailsNavigator } from "../features/payments/details/navigation/navigator";
+import { PaymentsMethodDetailsRoutes } from "../features/payments/details/navigation/routes";
+import { PaymentsOnboardingNavigator } from "../features/payments/onboarding/navigation/navigator";
+import { PaymentsOnboardingRoutes } from "../features/payments/onboarding/navigation/routes";
+import { PaymentsCheckoutNavigator } from "../features/payments/checkout/navigation/navigator";
+import { PaymentsCheckoutRoutes } from "../features/payments/checkout/navigation/routes";
+import { PaymentsTransactionNavigator } from "../features/payments/transaction/navigation/navigator";
+import { PaymentsTransactionRoutes } from "../features/payments/transaction/navigation/routes";
+import ServicesNavigator from "../features/services/navigation/navigator";
+import { SERVICES_ROUTES } from "../features/services/navigation/routes";
 import UADONATION_ROUTES from "../features/uaDonations/navigation/routes";
 import { UAWebViewScreen } from "../features/uaDonations/screens/UAWebViewScreen";
-import { WalletBarcodeNavigator } from "../features/payments/barcode/navigation/navigator";
-import { WalletBarcodeRoutes } from "../features/payments/barcode/navigation/routes";
-import {
-  WalletDetailsNavigator,
-  WalletDetailsRoutes
-} from "../features/payments/details/navigation/navigator";
-import {
-  WalletOnboardingNavigator,
-  WalletOnboardingRoutes
-} from "../features/payments/onboarding/navigation/navigator";
-import { WalletPaymentNavigator } from "../features/payments/payment/navigation/navigator";
-import { WalletPaymentRoutes } from "../features/payments/payment/navigation/routes";
-import {
-  WalletTransactionNavigator,
-  WalletTransactionRoutes
-} from "../features/payments/transaction/navigation/navigator";
 import { ZendeskStackNavigator } from "../features/zendesk/navigation/navigator";
 import ZENDESK_ROUTES from "../features/zendesk/navigation/routes";
 import { GalleryPermissionInstructionsScreen } from "../screens/misc/GalleryPermissionInstructionsScreen";
@@ -74,10 +72,6 @@ import {
   isIdPayEnabledSelector
 } from "../store/reducers/backendStatus";
 import { isGestureEnabled } from "../utils/navigation";
-import { MessagesStackNavigator } from "../features/messages/navigation/MessagesNavigator";
-import { MESSAGES_ROUTES } from "../features/messages/navigation/routes";
-import { SERVICES_ROUTES } from "../features/services/navigation/routes";
-import ServicesNavigator from "../features/services/navigation/navigator";
 import CheckEmailNavigator from "./CheckEmailNavigator";
 import OnboardingNavigator from "./OnboardingNavigator";
 import { AppParamsList } from "./params/AppParamsList";
@@ -294,36 +288,36 @@ const AuthenticatedStackNavigator = () => {
       )}
 
       <Stack.Screen
-        name={WalletOnboardingRoutes.WALLET_ONBOARDING_MAIN}
-        component={WalletOnboardingNavigator}
+        name={PaymentsOnboardingRoutes.PAYMENT_ONBOARDING_NAVIGATOR}
+        component={PaymentsOnboardingNavigator}
         options={{ gestureEnabled: isGestureEnabled, ...hideHeaderOptions }}
       />
       <Stack.Screen
-        name={WalletPaymentRoutes.WALLET_PAYMENT_MAIN}
-        component={WalletPaymentNavigator}
+        name={PaymentsCheckoutRoutes.PAYMENT_CHECKOUT_NAVIGATOR}
+        component={PaymentsCheckoutNavigator}
         options={{
           gestureEnabled: isGestureEnabled,
           ...hideHeaderOptions
         }}
       />
       <Stack.Screen
-        name={WalletDetailsRoutes.WALLET_DETAILS_MAIN}
-        component={WalletDetailsNavigator}
+        name={PaymentsMethodDetailsRoutes.PAYMENT_METHOD_DETAILS_NAVIGATOR}
+        component={PaymentsMethodDetailsNavigator}
         options={{
           gestureEnabled: isGestureEnabled,
           ...hideHeaderOptions
         }}
       />
       <Stack.Screen
-        name={WalletTransactionRoutes.WALLET_TRANSACTION_MAIN}
-        component={WalletTransactionNavigator}
+        name={PaymentsTransactionRoutes.PAYMENT_TRANSACTION_NAVIGATOR}
+        component={PaymentsTransactionNavigator}
         options={{
           gestureEnabled: isGestureEnabled,
           ...hideHeaderOptions
         }}
       />
       <Stack.Screen
-        name={WalletBarcodeRoutes.WALLET_BARCODE_MAIN}
+        name={PaymentsBarcodeRoutes.PAYMENT_BARCODE_NAVIGATOR}
         component={WalletBarcodeNavigator}
         options={{
           ...TransitionPresets.ModalSlideFromBottomIOS,